requirements.txt
# Python >= 3.10 required (developed with 3.13)
sentence-transformers
chromadb
litellm
python-dotenv
